test: split Obsidian integration harness
Group real-runtime checks by plugin, provider, vault, and chat while keeping orchestration and execution order in the harness.
