+++ test/obsidian/checks/assert.ts
@@ -1,0 +1,3 @@
+export function assert(condition: unknown, message: string): asserts condition {
+	if (!condition) throw new Error(message)
+}
