+++ test/obsidian/checks/vault.ts
@@ -1,0 +1,134 @@
+import type { App } from 'obsidian'
+import { assert } from './assert'
+
+const E2E_ROOT = '.agents/nutstore-sync/e2e'
+
+export async function roundTripsVaultAdapterContent(app: App) {
+	const path = `${E2E_ROOT}/中性 sample 🌱.txt`
+	const content = 'neutral content / 中性内容 🌱'
+	await app.vault.adapter.mkdir(E2E_ROOT)
+	await app.vault.adapter.write(path, content)
+	assert(
+		(await app.vault.adapter.read(path)) === content,
+		'Vault adapter did not round-trip Unicode content',
+	)
+	await app.vault.adapter.remove(path)
+	assert(
+		!(await app.vault.adapter.exists(path)),
+		'Vault adapter did not remove its test file',
+	)
+}
+
+export async function preservesBashHeredocUtf8(app: App) {
+	const { execVaultBash } = await import('~/ai/tools/bash/runtime')
+	const path = `${E2E_ROOT}/中性 heredoc 🌱.md`
+	const content = [
+		'---',
+		'tags:',
+		'  - 中性标签 🌱',
+		'---',
+		'',
+		'# 中性标题',
+		'',
+		'正文：中文、English、Emoji 🛠️',
+	].join('\n')
+	const result = await execVaultBash(
+		app,
+		[`cat > "/${path}" << 'ENDOFFILE'`, content, 'ENDOFFILE'].join('\n'),
+	)
+	assert(result.exitCode === 0, 'Bash heredoc write failed')
+	const bytes = await app.vault.adapter.readBinary(path)
+	assert(
+		new TextDecoder().decode(bytes) === `${content}\n`,
+		'Bash heredoc write did not preserve UTF-8 content',
+	)
+	// The MSB guest is disposable; keep the file available for failure
+	// diagnostics instead of racing Obsidian's external-file watcher.
+}
+
+export async function resolvesResourceDataUrls(app: App) {
+	const { resolveResourceDataUrl } =
+		await import('~/ai/tools/resource-data-url')
+	const path = `${E2E_ROOT}/中性 resource 🌱.txt`
+	const content = 'neutral resource / 中性资源 🌱'
+	await app.vault.adapter.write(path, content)
+	try {
+		const result = await resolveResourceDataUrl(app, `/${path}`, 'text/plain')
+		assert(result, 'Resource data URL was not resolved from the Vault')
+		const encoded = result.slice(result.indexOf(',') + 1)
+		const bytes = Uint8Array.from(atob(encoded), (character) =>
+			character.charCodeAt(0),
+		)
+		assert(
+			new TextDecoder().decode(bytes) === content,
+			'Resource data URL did not preserve Unicode content',
+		)
+	} finally {
+		await app.vault.adapter.remove(path)
+	}
+}
+
+export async function skipsStaleVaultSkillEntries(app: App) {
+	const { SkillRepository, SKILLS_ROOT } =
+		await import('~/ai/skills/repository')
+	const { ObsidianVaultFs } = await import('~/ai/tools/bash/fs')
+	const adapter = app.vault.adapter
+	const stableFolder = `${SKILLS_ROOT}/steady-skill`
+	const stableSkill = `${stableFolder}/SKILL.md`
+	const looseFile = `${SKILLS_ROOT}/中性文件🌱.md`
+	const staleFolder = `${SKILLS_ROOT}/temporary-中性-🌱`
+	const staleSkill = `${staleFolder}/SKILL.md`
+
+	if (!(await adapter.exists(SKILLS_ROOT))) await adapter.mkdir(SKILLS_ROOT)
+	await adapter.mkdir(stableFolder)
+	await adapter.write(
+		stableSkill,
+		'---\nname: steady-skill\ndescription: Neutral Skill / 中性技能 🌱\n---\n',
+	)
+	await adapter.write(looseFile, 'neutral loose file / 中性普通文件 🌱')
+
+	const originalList = adapter.list
+	const originalStat = adapter.stat
+	adapter.list = async (path) => {
+		const listed = await originalList.call(adapter, path)
+		return path === SKILLS_ROOT
+			? { ...listed, folders: [...listed.folders, staleFolder] }
+			: listed
+	}
+	adapter.stat = async (path) => {
+		if (path === staleFolder || path === staleSkill) {
+			throw new Error(`ENOENT: neutral unavailable entry '${path}'`)
+		}
+		return originalStat.call(adapter, path)
+	}
+
+	try {
+		const repository = new SkillRepository(app, [])
+		await repository.refresh()
+		assert(
+			repository.getCatalog().length === 1 &&
+				repository.getCatalog()[0].name === 'steady-skill',
+			'Stable Vault Skill was not discovered',
+		)
+		assert(
+			repository
+				.discover()
+				.diagnostics.some((diagnostic) => diagnostic.path === staleSkill),
+			'Stale Vault Skill did not produce a diagnostic',
+		)
+
+		const fs = new ObsidianVaultFs(app.vault)
+		const entries = await fs.readdir(`/${SKILLS_ROOT}`)
+		assert(
+			entries.includes('steady-skill') && entries.includes('中性文件🌱.md'),
+			'Stable Skill entries were hidden from Bash',
+		)
+		assert(
+			!entries.includes('temporary-中性-🌱'),
+			'Stale Skill entry was exposed to Bash',
+		)
+	} finally {
+		adapter.list = originalList
+		adapter.stat = originalStat
+	}
+}
